{"types":"C","runs":5}

+++ Problem30.java
@@ -3,7 +3,7 @@ import acm.program.ConsoleProgram;
 public class Problem30 extends ConsoleProgram {
 	public void run(){
 	double num = readInt("Give me a number");
-	for(int i = 1; Math.pow(10,i)<num; i++){
+	for(double i = 1; Math.pow(10,i)<num; i++){
 	double sub = num/(Math.pow(10,i));
 	double min = num-sub*Math.pow(10,i);
 	println(min);
